+++ v2023/src/stars/mod.rs
@@ -2,3 +2,4 @@ use super::gifts;
 
 pub mod day_01;
 pub mod day_02;
+pub mod day_03;