+++ CHANGELOG.md
@@ -5,6 +5,10 @@ all notable changes to this project will be documented here.
 the format is inspired by [keep a changelog][changelog] with entries logged as
 [conventional commits][commits] and marked with [calendar versioning][calver].
 
+## [unreleased]
+
+- feat(2023.3.1): compute the sum of parts for an engine schematic
+
 ## [0.2023.5]
 
 - docs(release): tag this release as today comes to an eventful end